+++ ccc-20200701/ChangeLog.md
@@ -1,3 +1,10 @@
+2024-09-07 Version: 2.29.2
+- Update API ListCallDetailRecordsV2: add param ReleaseInitiatorList.
+- Update API ListCallDetailRecordsV2: add param ReleaseReasonList.
+- Update API ListCallDetailRecordsV2: update response param.
+- Update API MakeCall: add param FlashSmsVariables.
+
+
 2024-08-13 Version: 2.29.1
 - Update API LaunchSurvey: update param ContactFlowId.